backend/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlmodel==0.0.14
sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0  # PostgreSQL async driver

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Authentication and Security
better-exceptions==0.3.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and Environment
python-dotenv==1.0.0

# CORS and Middleware
python-dateutil==2.8.2

# Type Hints
typing-extensions==4.8.0
